// client/src/protobuf/gen/user_inputs.ts
// @generated by protobuf-ts 2.11.1
// @generated from protobuf file "user_inputs.proto" (syntax proto3)
// tslint:disable
import type { BinaryWriteOptions } from "@protobuf-ts/runtime";
import type { IBinaryWriter } from "@protobuf-ts/runtime";
import { WireType } from "@protobuf-ts/runtime";
import type { BinaryReadOptions } from "@protobuf-ts/runtime";
import type { IBinaryReader } from "@protobuf-ts/runtime";
import { UnknownFieldHandler } from "@protobuf-ts/runtime";
import type { PartialMessage } from "@protobuf-ts/runtime";
import { reflectionMergePartial } from "@protobuf-ts/runtime";
import { MessageType } from "@protobuf-ts/runtime";
/**
 * @generated from protobuf message UserInputMessage
 */
export interface UserInputMessage {
    /**
     * @generated from protobuf field: int32 incomingPacketType = 1
     */
    incomingPacketType: number;
    /**
     * @generated from protobuf field: int32 direction = 2
     */
    direction: number;
}
// @generated message type with reflection information, may provide speed optimized methods
class UserInputMessage$Type extends MessageType<UserInputMessage> {
    constructor() {
        super("UserInputMessage", [
            { no: 1, name: "incomingPacketType", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "direction", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<UserInputMessage>): UserInputMessage {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.incomingPacketType = 0;
        message.direction = 0;
        if (value !== undefined)
            reflectionMergePartial<UserInputMessage>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: UserInputMessage): UserInputMessage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 incomingPacketType */ 1:
                    message.incomingPacketType = reader.int32();
                    break;
                case /* int32 direction */ 2:
                    message.direction = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: UserInputMessage, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 incomingPacketType = 1; */
        if (message.incomingPacketType !== 0)
            writer.tag(1, WireType.Varint).int32(message.incomingPacketType);
        /* int32 direction = 2; */
        if (message.direction !== 0)
            writer.tag(2, WireType.Varint).int32(message.direction);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message UserInputMessage
 */
export const UserInputMessage = new UserInputMessage$Type();