// client/src/protobuf/gen/wpm_packet.ts
// @generated by protobuf-ts 2.11.1
// @generated from protobuf file "wpm_packet.proto" (syntax proto3)
// tslint:disable
import type { BinaryWriteOptions } from "@protobuf-ts/runtime";
import type { IBinaryWriter } from "@protobuf-ts/runtime";
import { WireType } from "@protobuf-ts/runtime";
import type { BinaryReadOptions } from "@protobuf-ts/runtime";
import type { IBinaryReader } from "@protobuf-ts/runtime";
import { UnknownFieldHandler } from "@protobuf-ts/runtime";
import type { PartialMessage } from "@protobuf-ts/runtime";
import { reflectionMergePartial } from "@protobuf-ts/runtime";
import { MessageType } from "@protobuf-ts/runtime";
/**
 * @generated from protobuf message WpmPacket
 */
export interface WpmPacket {
    /**
     * @generated from protobuf field: int32 packetType = 1
     */
    packetType: number;
    /**
     * @generated from protobuf field: int32 payload = 2
     */
    payload: number;
}
// @generated message type with reflection information, may provide speed optimized methods
class WpmPacket$Type extends MessageType<WpmPacket> {
    constructor() {
        super("WpmPacket", [
            { no: 1, name: "packetType", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "payload", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<WpmPacket>): WpmPacket {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.packetType = 0;
        message.payload = 0;
        if (value !== undefined)
            reflectionMergePartial<WpmPacket>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: WpmPacket): WpmPacket {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 packetType */ 1:
                    message.packetType = reader.int32();
                    break;
                case /* int32 payload */ 2:
                    message.payload = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: WpmPacket, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 packetType = 1; */
        if (message.packetType !== 0)
            writer.tag(1, WireType.Varint).int32(message.packetType);
        /* int32 payload = 2; */
        if (message.payload !== 0)
            writer.tag(2, WireType.Varint).int32(message.payload);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message WpmPacket
 */
export const WpmPacket = new WpmPacket$Type();